+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionUtils.kt
@@ -231,6 +231,7 @@ object ActionUtils {
         ActionId.PERFORM_IME_ACTION -> ActionCategory.KEYBOARD
         ActionId.SWITCH_KEYBOARD -> ActionCategory.KEYBOARD
         ActionId.CYCLE_KEYBOARD_LANGUAGE -> ActionCategory.KEYBOARD
+        ActionId.CYCLE_KEYBOARD -> ActionCategory.KEYBOARD
         ActionId.LOCK_DEVICE -> ActionCategory.INTERFACE
         ActionId.POWER_ON_OFF_DEVICE -> ActionCategory.INTERFACE
         ActionId.SECURE_LOCK_DEVICE -> ActionCategory.INTERFACE
@@ -438,6 +439,8 @@ object ActionUtils {
         ActionId.SWITCH_KEYBOARD -> R.string.action_switch_keyboard
 
         ActionId.CYCLE_KEYBOARD_LANGUAGE -> R.string.action_cycle_keyboard_language
+
+        ActionId.CYCLE_KEYBOARD -> R.string.action_cycle_keyboard
 
         ActionId.TOGGLE_AIRPLANE_MODE -> R.string.action_toggle_airplane_mode
 
@@ -912,6 +915,10 @@ object ActionUtils {
 
             ActionId.CYCLE_KEYBOARD_LANGUAGE -> return listOf(Permission.WRITE_SECURE_SETTINGS)
 
+            ActionId.CYCLE_KEYBOARD -> if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
+                return listOf(Permission.WRITE_SECURE_SETTINGS)
+            }
+
             ActionId.TOGGLE_AIRPLANE_MODE,
             ActionId.ENABLE_AIRPLANE_MODE,
             ActionId.DISABLE_AIRPLANE_MODE,
@@ -1062,6 +1069,7 @@ object ActionUtils {
         ActionId.PERFORM_IME_ACTION -> Icons.Outlined.Keyboard
         ActionId.SWITCH_KEYBOARD -> Icons.Outlined.Keyboard
         ActionId.CYCLE_KEYBOARD_LANGUAGE -> Icons.Outlined.Keyboard
+        ActionId.CYCLE_KEYBOARD -> Icons.Outlined.Keyboard
         ActionId.TOGGLE_AIRPLANE_MODE -> Icons.Outlined.AirplanemodeActive
         ActionId.ENABLE_AIRPLANE_MODE -> Icons.Outlined.AirplanemodeActive
         ActionId.DISABLE_AIRPLANE_MODE -> Icons.Outlined.AirplanemodeInactive